Refactor to make default layout less opininated
Original default layout assumed all pages had the same header configuration (same banner or not, and same choice of whether the navbar is above or below the branding element). This refactoring takes that decision away from the default layout, and puts it into the individual layouts that inherit from default.